Change altr#show() format to be reasonable one
The old format "$current -> $forward -> $back" doesn't make sense if
a rule is defined with more than 3 patterns such as the following:

        call altr#define('a', 'b', 'c', 'd')

altr#show('a') printed "'a' -> 'b' -> 'd'", but the next pattern of 'b'
is 'c', not 'd'.

The new altr#show() prints "'d' <- 'a' -> 'b'", and it makes sense.