Fix altr#_switch() to distinguish similar paths
altr#_switch() uses :buffer to switch to an already opened path, and
:buffer requires a buffer number to switch.  So that altr#_switch() uses
bufnr() to determine whether a path is already opened or not.

But bufnr() doesn't use a given {expr} literally.  It should be escaped
properly.  Otherwise, altr#_switch() switches to an unexpected buffer.

Fix this bug by escaping {expr} for bufnr() properly.