Fix no rule is matched though rule exists
Though rule exists, following message is printed
when current directory and buffer's file path are the same.

"altr: No rule is matched to the current buffer name."

This problem is caused to use bufname("%"). Because bufname() function
return 'relative path'.
So I change to use expand("%:p") instead of bufname("%").